requirements.txt
aiosqlite==0.17.0; python_version >= "3.7" and python_version < "4.0"
asyncpg==0.24.0; python_full_version >= "3.6.0"
click==7.1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
ecdsa==0.17.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
fastapi==0.68.1; python_version >= "3.6"
h11==0.12.0; python_version >= "3.6"
iso8601==0.1.16; python_version >= "3.7" and python_version < "4.0"
passlib==1.7.4
pyasn1==0.4.8; python_version >= "3.5" and python_version < "4"
pydantic==1.8.2; python_full_version >= "3.6.1" and python_version >= "3.6"
pypika-tortoise==0.1.1; python_version >= "3.7" and python_version < "4.0"
python-jose==3.3.0
pytz==2021.3; python_version >= "3.7" and python_version < "4.0"
rsa==4.7.2; python_version >= "3.5" and python_version < "4"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
starlette==0.14.2; python_version >= "3.6"
tortoise-orm==0.17.7; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.10.0.2; python_full_version >= "3.6.1" and python_version >= "3.7" and python_version < "4.0"
uvicorn==0.13.4